Protect box endpoints

Closes #60

See merge request ase-22-23/team40/ase-project-2022-team-40!35

+++ microservices/delivery-service/src/main/java/com/group40/deliveryservice/controller/BoxController.java
@@ -43,38 +43,83 @@ public class BoxController {
 
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
-    public List<BoxResponse> getAllBoxes() {
-        return boxService.getAllBoxes();
+    public ResponseEntity<?> getAllBoxes(@RequestHeader(HttpHeaders.AUTHORIZATION) String token) throws JSONException, IOException {
+        User user = userService.getUser(token);
+        if (user.getRole().equals(ERole.ROLE_DISPATCHER)) {
+            return ResponseEntity.ok(boxService.getAllBoxes());
+        }else if (user.getRole().equals(ERole.ROLE_DELIVERER)){
+            return ResponseEntity.ok(boxService.getBoxesByDeliverer(user.getId()));
+        }else if (user.getRole().equals(ERole.ROLE_CUSTOMER)){
+            return ResponseEntity.ok(boxService.getBoxesByCustomer(user.getId()));
+        } else{
+            return ResponseEntity.badRequest().body("Not authorized!");
+        }
     }
 
 
     @GetMapping("/box")
     @ResponseStatus(HttpStatus.OK)
-    public BoxResponse getBox(@RequestParam String id) throws Exception  { return boxService.getBox(id);}
+    public ResponseEntity<?> getBox(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @RequestParam String id) throws Exception {
+        User user = userService.getUser(token);
+        BoxResponse box = boxService.getBox(id);
+        if (user.getRole().equals(ERole.ROLE_DISPATCHER) ||
+                (box.getAssignedTo().equals(user.getId()) && user.getRole().equals(ERole.ROLE_DELIVERER)) ||
+                (box.getAssignedCustomers().contains(user.getId()) && user.getRole().equals(ERole.ROLE_CUSTOMER))) {
+            return ResponseEntity.ok(box);
+        } else {
+            return ResponseEntity.badRequest().body("Not authorized!");
+        }
+
+    }
 
     @PostMapping("/box")
     @ResponseStatus(HttpStatus.OK)
-    public BoxResponse updateBox(@RequestParam String id , @RequestBody Map<String, String> obj) throws Exception {
-        return boxService.updateBox(id, obj);
+    public ResponseEntity<?> updateBox(@RequestHeader(HttpHeaders.AUTHORIZATION) String token,
+                                       @RequestParam String id, @RequestBody Map<String, String> obj) throws Exception {
+        User user = userService.getUser(token);
+        if (user.getRole().equals(ERole.ROLE_DISPATCHER)) {
+            return ResponseEntity.ok(boxService.updateBox(id, obj));
+        } else {
+            return ResponseEntity.badRequest().body("Not authorized!");
+        }
     }
 
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
-    Box replaceBox(@RequestBody Box newBox, @PathVariable String id) {
-        return boxService.replaceBox(newBox, id);
+    ResponseEntity<?> replaceBox(@RequestHeader(HttpHeaders.AUTHORIZATION) String token,
+                   @RequestBody Box newBox, @PathVariable String id) throws JSONException, IOException {
+        User user = userService.getUser(token);
+        if (user.getRole().equals(ERole.ROLE_DISPATCHER)) {
+            return ResponseEntity.ok(boxService.replaceBox(newBox, id));
+        } else {
+            return ResponseEntity.badRequest().body("Not authorized!");
+        }
     }
 
     @GetMapping("/deliverer/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public List<BoxResponse> getBoxesByDeliverer(@PathVariable String id) {
-        return boxService.getBoxesByDeliverer(id);
+    public ResponseEntity<?> getBoxesByDeliverer(@RequestHeader(HttpHeaders.AUTHORIZATION) String token,
+                                                 @PathVariable String id) throws JSONException, IOException {
+        User user = userService.getUser(token);
+        if (user.getRole().equals(ERole.ROLE_DELIVERER)) {
+            return ResponseEntity.ok(boxService.getBoxesByDeliverer(id));
+        } else {
+            return ResponseEntity.badRequest().body("Not authorized!");
+        }
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public String deleteBox(@PathVariable String id) throws Exception {
-        boxService.deleteBox(id);
-        return "{}";
+    public ResponseEntity<?> deleteBox(@RequestHeader(HttpHeaders.AUTHORIZATION) String token,
+                                       @PathVariable String id) throws Exception {
+        User user = userService.getUser(token);
+        if (user.getRole().equals(ERole.ROLE_DISPATCHER)) {
+            boxService.deleteBox(id);
+            return ResponseEntity.ok("Box deleted!");
+        } else {
+            return ResponseEntity.badRequest().body("Not authorized!");
+        }
+
     }
 
     @GetMapping("/current")
